Common CreatePolicy
