+++ portals/publisher/src/main/webapp/source/src/app/components/GlobalPolicies/Policies/CreatePolicy.tsx
@@ -17,130 +17,12 @@
  */
 
 import React, { useContext, useState } from 'react';
-import { Typography, makeStyles, Theme } from '@material-ui/core';
-import Box from '@material-ui/core/Box';
-import { Link } from 'react-router-dom';
-import Dialog from '@material-ui/core/Dialog';
-import DialogContent from '@material-ui/core/DialogContent';
-import DialogContentText from '@material-ui/core/DialogContentText';
-import Icon from '@material-ui/core/Icon';
-import IconButton from '@material-ui/core/IconButton';
-import LaunchIcon from '@material-ui/icons/Launch';
-import { FormattedMessage } from 'react-intl';
 import API from 'AppData/api.js';
 import Alert from 'AppComponents/Shared/Alert';
 import ApiContext from 'AppComponents/Apis/Details/components/ApiContext';
-import CONSTS from 'AppData/Constants';
+import CreatePolicyShared from 'AppComponents/Shared/PoliciesUI/CreatePolicy';
 import type { CreatePolicySpec } from './Types';
 import PolicyCreateForm from './PolicyForm/PolicyCreateForm';
-
-/** Shared UI Component */
-const useStyles = makeStyles((theme: Theme) => ({
-    link: {
-        color: theme.palette.primary.dark,
-        marginLeft: theme.spacing(2),
-        display: 'inline',
-    },
-}));
-
-interface CreatePolicySharedProps {
-    handleDialogClose: () => void;
-    dialogOpen: boolean;
-    stopPropagation: any;
-    onSave: (policySpecification: CreatePolicySpec) => void;
-    synapsePolicyDefinitionFile: any[];
-    setSynapsePolicyDefinitionFile: any;
-    ccPolicyDefinitionFile: any[];
-    setCcPolicyDefinitionFile: any;
-    saving: any;
-}
-
-const CreatePolicyShared: React.FC<CreatePolicySharedProps> = ({
-    handleDialogClose,
-    dialogOpen,
-    stopPropagation,
-    onSave,
-    synapsePolicyDefinitionFile,
-    setSynapsePolicyDefinitionFile,
-    ccPolicyDefinitionFile,
-    setCcPolicyDefinitionFile,
-    saving
-}) => {
-    const classes = useStyles();
-
-    return (
-        <>
-            <Dialog
-                maxWidth='md'
-                open={dialogOpen}
-                aria-labelledby='form-dialog-title'
-                onClose={handleDialogClose}
-                onClick={stopPropagation}
-                fullWidth
-            >
-                <Box
-                    display='flex'
-                    justifyContent='space-between'
-                    alignItems='center'
-                    flexDirection='row'
-                    px={3}
-                    pt={3}
-                >
-                    <Box display='flex'>
-                        <Typography variant='h4' component='h2'>
-                            <FormattedMessage
-                                id='Apis.Details.Policies.CreatePolicy.create.new.policy'
-                                defaultMessage='Create New Policy'
-                            />
-                        </Typography>
-                    </Box>
-                    <Box display='flex'>
-                        <IconButton
-                            color='inherit'
-                            onClick={handleDialogClose}
-                            aria-label='Close'
-                        >
-                            <Icon>close</Icon>
-                        </IconButton>
-                    </Box>
-                </Box>
-                <DialogContent>
-                    <Box my={2}>
-                        <DialogContentText>
-                            <PolicyCreateForm
-                                onSave={onSave}
-                                synapsePolicyDefinitionFile={synapsePolicyDefinitionFile}
-                                setSynapsePolicyDefinitionFile={setSynapsePolicyDefinitionFile}
-                                ccPolicyDefinitionFile={ccPolicyDefinitionFile}
-                                setCcPolicyDefinitionFile={setCcPolicyDefinitionFile}
-                                onCancel={handleDialogClose}
-                                saving={saving}
-                            />
-                        </DialogContentText>
-                    </Box>
-                </DialogContent>
-                <Box
-                    display='flex'
-                    flexDirection='row'
-                    justifyContent='right'
-                    px={3}
-                    pb={3}
-                >
-                    <Link to={CONSTS.PATH_TEMPLATES.COMMON_POLICIES}>
-                        <Typography className={classes.link} variant='caption'>
-                            Want to create a common policy that will be visible to all APIs instead?
-                            <LaunchIcon
-                                style={{ marginLeft: '2px' }}
-                                fontSize='small'
-                            />
-                        </Typography>
-                    </Link>
-                </Box>
-            </Dialog>
-        </>
-    );
-}
-/** Shared UI Component Ends */
 
 interface CreatePolicyProps {
     handleDialogClose: () => void;
@@ -221,6 +103,7 @@ const CreatePolicy: React.FC<CreatePolicyProps> = ({
             ccPolicyDefinitionFile={ccPolicyDefinitionFile}
             setCcPolicyDefinitionFile={setCcPolicyDefinitionFile}
             saving={saving}
+            PolicyCreateForm={PolicyCreateForm}
         />
     );
 };
